move parse tree node class out to its own module
